French version / Correction de bug : -Corrige le bug de mauvaise prise en compte du redimensionnement de la fenêtre du navigateur de fichier (certaines vignettes ne sont plus sélectionnable après redimensionnement) -Correction de plus de 70 phrases anglaises incorrectes dans l'interface (orthographe ou grammaire) - Merci à J.G. Visser pour son aide -Corrige les problèmes d'accent sous Windows pour les répertoires favoris et les répertoires sauvegardés (dernier utilisé pour le navigateur, la génération, etc...)
English version / Bug fix :
-Bug fix when resize file browser window (some thumbnails was selectable after resize)
-Correction of more than 70 incorrect English sentences in the interface (spelling or grammar) - Thanks to J.G. Visser for his help
-Correct the problems of accent under Windows for the favourite directories and the saved directories (the last one used for the browser, the generation, etc.)

French version / Nouvelles fonctionnalités :
-Ajoute le raccourcie clavier F2 pour le renomage des fichiers dans le navigateur de fichier
-Réduit la qualité des vidéos dans la prévisualisation pour accélérer l'affichage
-Prend en charge les vidéos au format AVCHD (.mts) - expérimental
-Affiche la position de fin des plans, sur les vignettes de plan, dans la boite de dialogue d'edition des diapositives
-Rend l'option "Afficher des informations sur les blocs" active même si le bloc est caché, dans la boite d'edition des diapositives
-Supprime le raccourcie clavier F5 vers l’édition du fond d'écran pour ne pas faire doublon avec le raccourcie de rafraîchissement du navigateur

English version / New features :
-Add F2 shortcut to rename file in file browser window
-Reduce video quality during preview to speed up display
-Enable use of AVCHD video (.mts) - experimental
-Display shots end position on shot thumbnails in edit slide dialog
-Set "Display block properties" active even if block is hiden in slide edit dialog
-Remove shortcut F5 to edit background to not duplicate it with shortuc ot refresh browser

git-svn-id: file:///media/werkstatt/064f12d4-5f27-4839-9537-9ffb7939d236/ffDiaporama-Entwicklung-2023/ffdiaporama-svn/trunk@317 f803d417-eb43-4c80-8848-1a7435ce69a7

+++ ffDiaporama/src/engine/cDeviceModelDef.cpp
@@ -536,19 +536,21 @@ bool cDeviceModelList::Initffmpeg(QString &BinaryEncoderPath) {
         QProcess    Process;
 
         #ifdef LIBAV_AVCONV
-        // Search avconv
-        BinaryEncoderPath="avconv";
-        Process.start(BinaryEncoderPath,QString("-version").split(";"));
-        if (Process.waitForStarted(-1)) {
-            if (!Process.waitForFinished()) Process.kill();
-                else Continue=true;
+        if (!Continue) {
+            // Search avconv
+            BinaryEncoderPath="avconv";
+            Process.start(BinaryEncoderPath,QString("-version").split(";"));
+            if (Process.waitForStarted(-1)) {
+                if (!Process.waitForFinished()) Process.kill();
+                    else Continue=true;
+            }
+            ToLog(LOGMSG_INFORMATION,QString("Try to found %1 ... %2").arg(BinaryEncoderPath).arg(Continue?"found":"not found"));
         }
-        ToLog(LOGMSG_INFORMATION,QString("Try to found %1 ... %2").arg(BinaryEncoderPath).arg(Continue?"found":"not found"));
         #endif
 
         #ifdef LIBAV_FFMPEG
+        // Search ffmpeg
         if (!Continue) {
-            // Search ffmpeg
             BinaryEncoderPath="ffmpeg";
             Process.start(BinaryEncoderPath,QString("-version").split(";"));
             if (Process.waitForStarted(-1)) {
